refactor(mso_mdoc): split verifier.py into trust_store, cred_verifier, and pres_verifier modules
verifier.py was ~839 lines. Split into three focused modules:

- trust_store.py: TrustStore protocol + WalletTrustStore
- cred_verifier.py: MsoMdocCredVerifier + parsing helpers
  (PreverifiedMdocClaims, _parse_string_credential, _extract_mdoc_claims)
- pres_verifier.py: MsoMdocPresVerifier + OID4VP helpers + mdoc_verify
  (extract_mdoc_item_value, extract_verified_claims, MdocVerifyResult)

verifier.py is now a thin re-exporter for backward compatibility.

Update all test patch targets to reference the new module paths:
- mso_mdoc.mdoc.verifier.isomdl_uniffi -> cred_verifier / pres_verifier
- mso_mdoc.mdoc.verifier.Config -> pres_verifier
- mso_mdoc.mdoc.verifier.retrieve_or_create_did_jwk -> pres_verifier
- mso_mdoc.mdoc.verifier.MdocStorageManager -> trust_store

Tests that exercise mdoc_verify() also patch cred_verifier.isomdl_uniffi
since _parse_string_credential lives there and holds its own module reference.
