+++ oid4vc/mso_mdoc/tests/test_empty_trust_anchors.py
@@ -112,7 +112,7 @@ class TestEmptyTrustAnchorsCredVerifier:
         profile, _ = _make_profile()
         mock_mdoc = _make_mock_mdoc(verified=True)  # Rust "accepts" without trust anchors
 
-        with patch("mso_mdoc.mdoc.verifier.isomdl_uniffi") as mock_iso:
+        with patch("mso_mdoc.mdoc.cred_verifier.isomdl_uniffi") as mock_iso:
             mock_iso.MdocVerificationError = _iso_stub.MdocVerificationError
             mock_iso.Mdoc.from_string.return_value = mock_mdoc
 
@@ -148,7 +148,7 @@ class TestEmptyTrustAnchorsCredVerifier:
         profile, _ = _make_profile()
         mock_mdoc = _make_mock_mdoc(verified=True)
 
-        with patch("mso_mdoc.mdoc.verifier.isomdl_uniffi") as mock_iso:
+        with patch("mso_mdoc.mdoc.cred_verifier.isomdl_uniffi") as mock_iso:
             mock_iso.MdocVerificationError = _iso_stub.MdocVerificationError
             mock_iso.Mdoc.from_string.return_value = mock_mdoc
 
@@ -182,7 +182,7 @@ class TestEmptyTrustAnchorsCredVerifier:
         profile, _ = _make_profile()
         mock_mdoc = _make_mock_mdoc(verified=True, common_name="Test Root CA")
 
-        with patch("mso_mdoc.mdoc.verifier.isomdl_uniffi") as mock_iso:
+        with patch("mso_mdoc.mdoc.cred_verifier.isomdl_uniffi") as mock_iso:
             mock_iso.MdocVerificationError = _iso_stub.MdocVerificationError
             mock_iso.Mdoc.from_string.return_value = mock_mdoc
 
@@ -215,7 +215,7 @@ class TestEmptyTrustAnchorsCredVerifier:
         profile, _ = _make_profile()
         mock_mdoc = _make_mock_mdoc(verified=False)
 
-        with patch("mso_mdoc.mdoc.verifier.isomdl_uniffi") as mock_iso:
+        with patch("mso_mdoc.mdoc.cred_verifier.isomdl_uniffi") as mock_iso:
             mock_iso.MdocVerificationError = _iso_stub.MdocVerificationError
             mock_iso.Mdoc.from_string.return_value = mock_mdoc
 
@@ -241,9 +241,13 @@ class TestMdocVerifyEmptyTrustAnchors:
         """
         mock_mdoc = _make_mock_mdoc(verified=True)
 
-        with patch("mso_mdoc.mdoc.verifier.isomdl_uniffi") as mock_iso:
+        with patch("mso_mdoc.mdoc.pres_verifier.isomdl_uniffi") as mock_iso, patch(
+            "mso_mdoc.mdoc.cred_verifier.isomdl_uniffi"
+        ) as mock_iso_cred:
             mock_iso.MdocVerificationError = _iso_stub.MdocVerificationError
-            mock_iso.Mdoc.from_string.return_value = mock_mdoc
+            mock_iso_cred.MdocVerificationError = _iso_stub.MdocVerificationError
+            mock_iso_cred.Mdoc.from_string.return_value = mock_mdoc
+            mock_iso_cred.Mdoc.new_from_base64url_encoded_issuer_signed.return_value = mock_mdoc
 
             result = mdoc_verify("a0b1c2d3e4f5", trust_anchors=None)
 
@@ -256,9 +260,13 @@ class TestMdocVerifyEmptyTrustAnchors:
         """mdoc_verify(mso_mdoc, trust_anchors=[]) must return verified=False."""
         mock_mdoc = _make_mock_mdoc(verified=True)
 
-        with patch("mso_mdoc.mdoc.verifier.isomdl_uniffi") as mock_iso:
+        with patch("mso_mdoc.mdoc.pres_verifier.isomdl_uniffi") as mock_iso, patch(
+            "mso_mdoc.mdoc.cred_verifier.isomdl_uniffi"
+        ) as mock_iso_cred:
             mock_iso.MdocVerificationError = _iso_stub.MdocVerificationError
-            mock_iso.Mdoc.from_string.return_value = mock_mdoc
+            mock_iso_cred.MdocVerificationError = _iso_stub.MdocVerificationError
+            mock_iso_cred.Mdoc.from_string.return_value = mock_mdoc
+            mock_iso_cred.Mdoc.new_from_base64url_encoded_issuer_signed.return_value = mock_mdoc
 
             result = mdoc_verify("a0b1c2d3e4f5", trust_anchors=[])
 
@@ -281,9 +289,13 @@ class TestMdocVerifyEmptyTrustAnchors:
         )
         mock_mdoc = _make_mock_mdoc(verified=True, common_name="My CA")
 
-        with patch("mso_mdoc.mdoc.verifier.isomdl_uniffi") as mock_iso:
+        with patch("mso_mdoc.mdoc.pres_verifier.isomdl_uniffi") as mock_iso, patch(
+            "mso_mdoc.mdoc.cred_verifier.isomdl_uniffi"
+        ) as mock_iso_cred:
             mock_iso.MdocVerificationError = _iso_stub.MdocVerificationError
-            mock_iso.Mdoc.from_string.return_value = mock_mdoc
+            mock_iso_cred.MdocVerificationError = _iso_stub.MdocVerificationError
+            mock_iso_cred.Mdoc.from_string.return_value = mock_mdoc
+            mock_iso_cred.Mdoc.new_from_base64url_encoded_issuer_signed.return_value = mock_mdoc
 
             result = mdoc_verify("a0b1c2d3e4f5", trust_anchors=[pem_cert])
 
@@ -292,10 +304,13 @@ class TestMdocVerifyEmptyTrustAnchors:
 
     def test_mdoc_verify_parse_failure_returns_not_verified(self):
         """Parsing failure always returns verified=False regardless of trust anchors."""
-        with patch("mso_mdoc.mdoc.verifier.isomdl_uniffi") as mock_iso:
+        with patch("mso_mdoc.mdoc.pres_verifier.isomdl_uniffi") as mock_iso, patch(
+            "mso_mdoc.mdoc.cred_verifier.isomdl_uniffi"
+        ) as mock_iso_cred:
             mock_iso.MdocVerificationError = _iso_stub.MdocVerificationError
-            mock_iso.Mdoc.from_string.side_effect = Exception("CBOR parse error")
-            mock_iso.Mdoc.new_from_base64url_encoded_issuer_signed.side_effect = (
+            mock_iso_cred.MdocVerificationError = _iso_stub.MdocVerificationError
+            mock_iso_cred.Mdoc.from_string.side_effect = Exception("CBOR parse error")
+            mock_iso_cred.Mdoc.new_from_base64url_encoded_issuer_signed.side_effect = (
                 Exception("base64 parse error")
             )
 
@@ -327,7 +342,7 @@ class TestWalletTrustStoreEmptyCache:
         verifier = MsoMdocCredVerifier(trust_store=trust_store)
         mock_mdoc = _make_mock_mdoc(verified=True)
 
-        with patch("mso_mdoc.mdoc.verifier.isomdl_uniffi") as mock_iso:
+        with patch("mso_mdoc.mdoc.cred_verifier.isomdl_uniffi") as mock_iso:
             mock_iso.MdocVerificationError = _iso_stub.MdocVerificationError
             mock_iso.Mdoc.from_string.return_value = mock_mdoc
 
@@ -352,7 +367,7 @@ class TestWalletTrustStoreEmptyCache:
         profile, mock_session = _make_profile()
 
         # Patch MdocStorageManager to return zero trust anchors
-        with patch("mso_mdoc.mdoc.verifier.MdocStorageManager") as MockStorage:
+        with patch("mso_mdoc.mdoc.trust_store.MdocStorageManager") as MockStorage:
             MockStorage.return_value.get_all_trust_anchor_pems = AsyncMock(
                 return_value=[]
             )
@@ -362,7 +377,7 @@ class TestWalletTrustStoreEmptyCache:
         verifier = MsoMdocCredVerifier(trust_store=trust_store)
         mock_mdoc = _make_mock_mdoc(verified=True)
 
-        with patch("mso_mdoc.mdoc.verifier.isomdl_uniffi") as mock_iso:
+        with patch("mso_mdoc.mdoc.cred_verifier.isomdl_uniffi") as mock_iso:
             mock_iso.MdocVerificationError = _iso_stub.MdocVerificationError
             mock_iso.Mdoc.from_string.return_value = mock_mdoc
 
@@ -390,9 +405,9 @@ class TestEmptyTrustAnchorsPresVerifier:
         pres_record = _make_presentation_record()
 
         with (
-            patch("mso_mdoc.mdoc.verifier.isomdl_uniffi") as mock_iso,
-            patch("mso_mdoc.mdoc.verifier.Config") as mock_config,
-            patch("mso_mdoc.mdoc.verifier.retrieve_or_create_did_jwk") as mock_jwk_fn,
+            patch("mso_mdoc.mdoc.pres_verifier.isomdl_uniffi") as mock_iso,
+            patch("mso_mdoc.mdoc.pres_verifier.Config") as mock_config,
+            patch("mso_mdoc.mdoc.pres_verifier.retrieve_or_create_did_jwk") as mock_jwk_fn,
         ):
             mock_config.from_settings.return_value.endpoint = "https://issuer.example"
             mock_jwk = MagicMock()
